requirements.txt
pandas
numpy
scikit-learn
imbalanced-learn
xgboost
streamlit
joblib
matplotlib
seaborn
plotly
openpyxl
scipy
nbformat
pyyaml
pydantic
python-dotenv
fastapi
uvicorn
pytest
pytest-cov
httpx
